+++ armada_command/command_ssh.py
@@ -56,6 +56,10 @@ def command_ssh(args):
         raise ValueError('No microservice name supplied.')
 
     instances = [i for i in armada_utils.get_matched_containers(microservice_name) if 'kv_index' not in i]
+
+    if args.local:
+        instances = [i for i in instances if armada_utils.is_local_container(i['ServiceID'])]
+
     instances_count = len(instances)
 
     if instances_count < 1:
@@ -102,6 +106,6 @@ def command_ssh(args):
             .format(**locals())
         ssh_args = shlex.split(remote_ssh_chunk)
         ssh_args.extend(('sudo', docker_command))
-        print("Connecting to {0} on host {1}...".format(instance['name'], ssh_host))
+        print("Connecting to {0} on host {1}...".format(instance['ServiceName'], ssh_host))
 
     os.execvp(ssh_args[0], ssh_args)